pyproject.toml
[project]
name = "excvish"
version = "0.3.0"
description = "Extension of Computer Vision, Image Segmentation, and Hyper-spectral image"
readme = "README.md"
authors = [
    { name = "Yusaku Mandai", email = "yusaku.mandai.02@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "spectral>=0.23.1",
    "ultralytics>=8.3.209",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "pydoclint>=0.8.3",
    "pytest>=9.0.2",
    "ruff>=0.9.7",
    "scipy-stubs>=1.16.0.2",
    "tox>=4.24.1",
    "tox-uv>=1.25.0",
    "ty>=0.0.1a13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 99
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstyle
    "D401",
]
ignore = [
    "E741" # ambiguous-variable-name
]

[tool.ruff.lint.pydocstyle]
convention = "google"
